requirements.txt
amqp==5.1.1
asgiref==3.7.2
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.7
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.4
greenlet==3.0.1
kombu==5.3.2
pipenv==2023.10.3
platformdirs==3.11.0
prompt-toolkit==3.0.39
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uuid==1.30
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.2.9
whitenoise==6.6.0
